Remove ObjectIntIdentityMap list replacement
With 2-3 years more JVM experience under my belt, this no longer
looks worthwhile to implement. The JIT is more likely to inline ArrayList.get
into the map getter, and then coalesce the bounds checks.

Additionally, this optimization forces a dependency on a 3rd party
library to get an array list that gives us access to the backing array.